Move the table converter hack to its own file ...
... and add comments
